add edit product, fixing order, and fixing database order and product ( integer to bigInteger )

+++ database/migrations/2023_02_18_055612_create_products_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('nama_product');
             $table->string('image')->nullable();
             $table->string('spesifikasi');
-            $table->integer('harga');
-            $table->integer('kuantitas');
+            $table->bigInteger('harga');
+            $table->bigInteger('kuantitas');
             $table->timestamps();
         });
     }
